requirements.txt
# --- Core AgentBeats (A2A) ---
a2a-sdk[http-server]>=0.3.22
uvicorn>=0.23.0
fastapi>=0.100.0
httpx>=0.24.0
python-dotenv>=1.0.0

# --- Dependencias de Servidor (Obligatorias para A2A 0.3.x) ---
starlette
sse-starlette
sqlalchemy

# --- Inteligencia (Gemini & LangChain) ---
langchain-google-genai>=1.0.0
langchain-core>=0.1.0

# --- Datos y Benchmark (CRM Arena) ---
datasets>=2.14.0
pandas>=2.0.0
python-dateutil>=2.8.2

# --- Utilidades ---
pydantic>=2.0.0
tomli>=2.0.0
huggingface-hub